literally the most ridiculous bug I have ever seen
ok here we go:
Last commit switched over the Nametable backing map implementation from TreeMap to HashMap.
TreeMap is always aware of the key and value types (in this case, Identifier and Stackable respectively).
Therefore, when it needs to compare keys when retrieving/inserting/moving values,
it calls key1.equals(key2). Because both key1 and key2 are typed as Identifier (generic type K in the actual source),
Java picks a correct equals() override to invoke.
In my case, all Stackable subtypes have an equals(Stackable) implemented (because Stackable is an interface and cannot override the Object inherited method equals(Object) ).
And because Identifier is an implementor of Stackable, in the TreeMap equality check, Java will find that equals(Stackable) method and call it.
Which is correct because it checks for the desired identifier name equality.

Now, for whatever reason, HashMap does at some point in its internal method chain forget the types of the objects.
Despite it being a generic interface with type information always readily available.
This means that by the time it does its equality checks (which, of course, are still required for many operations),
it will consider key1 and key2 to be Objects. And call Object.equals(Object).
Which I, unfortunately, hadn't implemented in Identifier.
So the default check is basically pointer equality, which of course fails, and no key-value mapping is ever found again.
Thanks, Java.
Now I know to always implement Object.equals(Object) as well, just as this commit does for Identifier, which fixes everything.
Sorry for breaking your stuff with the last commit.
I really didn't see this one coming.

Thank you for coming to my TED talk.